+++ src/main/java/domain/BlackJackGame.java
@@ -10,6 +10,9 @@ import domain.card.CardDeck;
 import domain.participant.Participant;
 import domain.participant.Participants;
 import domain.participant.PlayerName;
+import domain.result.DealerResults;
+import domain.result.DealerWinningStatus;
+import domain.result.PlayerResults;
 import domain.result.PlayerWinningStatus;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -72,7 +75,7 @@ public class BlackJackGame {
     }
 
     public PlayerRevenues calculateRevenue(PlayerBetMonies playerBetMonies) {
-        ParticipantsResult participantsResult = BlackJackResultCalculator.calculate(participants);
+        ParticipantsResult participantsResult = calculateGameResult();
         List<PlayerRevenue> playerRevenues = new ArrayList<>();
 
         for (PlayerWinningStatus playerWinningStatus : participantsResult.playerResults()
@@ -90,6 +93,22 @@ public class BlackJackGame {
             playerRevenues.add(addValue);
         }
         return new PlayerRevenues(playerRevenues);
+    }
+
+    public ParticipantsResult calculateGameResult() {
+        Participant dealer = participants.getDealer();
+        List<DealerWinningStatus> dealerResults = new ArrayList<>();
+        List<PlayerWinningStatus> playerResults = new ArrayList<>();
+        for (Participant player : participants.getPlayers()) {
+            BlackJackWinningStatus dealerBlackJackWinningStatus = BlackJackWinningStatus.calculateDealerResult(
+                    dealer.isBlackJack(),
+                    dealer.getTotalValue(),
+                    player.isBlackJack(),
+                    player.getTotalValue());
+            dealerResults.add(new DealerWinningStatus(player.getName(), dealerBlackJackWinningStatus));
+            playerResults.add(new PlayerWinningStatus(player.getName(), dealerBlackJackWinningStatus.reverse()));
+        }
+        return new ParticipantsResult(new DealerResults(dealerResults), new PlayerResults(playerResults));
     }
 
     public Participant getDealer() {
